requirements.txt
flask==3.0.0
flask-limiter==3.5.0
requests==2.31.0
beautifulsoup4==4.12.2
apscheduler==3.10.4
python-dotenv==1.0.0
filelock==3.13.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
